move eventbus to its own contracts file
